feat: add types to `ComboButton`
Convert `ComboButton` to Typescript and export newly added
interface `ComboButtonProps`.
